pyproject.toml
[tool.poetry]
name = "drf-auth"
version = "0.1.0"
description = ""
authors = ["will-ing <will-ing@github.com>"]

[tool.poetry.dependencies]
python = "^3.8, <3.9"
pylint = "^2.6.0"
autopep8 = "^1.5.4"
django = "^3.1.1"
djangorestframework_simplejwt = "^4.4.0"
psycopg2-binary = "^2.8.6"
gunicorn = "^20.0.4"
whitenoise = "^5.2.0"
django-cors-headers = "^3.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
